Split guildScheduledEvent types
