+++ packages/types/src/index.ts
@@ -7,6 +7,7 @@ export * from './discord/emojis.js'
 export * from './discord/entitlements.js'
 export * from './discord/gateway.js'
 export * from './discord/guilds.js'
+export * from './discord/guildScheduledEvents.js'
 export * from './discord/interactions.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
